Rename `SwapButton` to `SwitchButton` as Swap is a reserved namespace
